Refactor error handling and trip management in TripViewModel and Repository
- Update `TripRepository` and `TripViewModel` to use `AppError` status codes instead of generic types.
- Refactor `TripsScreen` and `CreateTripScreen` to observe and display error messages via a unified `status` property.
- Improve trip creation and update flow in `CreateTripScreen` by handling navigation upon success.
- Clean up unused synchronous DAO methods and optimize itinerary item fetching using flows.
- Add `loadTrip` functionality to `TripViewModel` for better state management in `ItineraryScreen`.

+++ app/app/src/main/java/com/monarca/smarttravel/domain/interfaces/TripRepository.kt
@@ -10,7 +10,7 @@ interface TripRepository {
     fun getAllTrips(): Flow<List<Trip>>
     fun getTripsByUser(userId: Int): Flow<List<Trip>>
     suspend fun getTripById(tripId: Int): Trip?
-    suspend fun addTrip(trip: Trip): Long
+    suspend fun addTrip(trip: Trip): Int
     suspend fun updateTrip(trip: Trip): Int
     suspend fun deleteTrip(tripId: Int): Int
     suspend fun updateImage(tripId: Int, newImageResId: Int?): Int
